requirements.txt
click
PyGithub
python-gitlab
loguru
